refactor: #1325 Remove old installations component
