+++ packages/marketplace/src/components/pages/app-detail/developer/developer-app-content.tsx
@@ -1,16 +1,16 @@
 import * as React from 'react'
+import dayjs from 'dayjs'
 import { Dispatch } from 'redux'
 import { GET_ALL_PAGE_SIZE } from '@/constants/paginator'
 import { appInstallationsRequestData } from '@/actions/app-installations'
 import { selectDeveloperId } from '@/selector'
-import { FlexContainerBasic } from '@reapit/elements'
+import { FlexContainerBasic, Button, DATE_TIME_FORMAT } from '@reapit/elements'
 import ConfirmUninstall from '@/components/ui/app-installations/confirm-uninstall'
 import { PagedResultInstallationModel_, InstallationModel } from '@reapit/foundations-ts-definitions'
 import { handleUninstall, handleAfterClose } from '@/components/ui/app-installations/app-installations-modal'
 import { useSelector, useDispatch } from 'react-redux'
 import { selectInstallationAppData } from '@/selector/installations'
 import { Modal } from '@reapit/elements'
-import { generateColumns } from '@/components/ui/app-installations/installations'
 import { AppDetailDataNotNull } from '@/reducers/client/app-detail'
 import { DeveloperAppDetailState } from '@/reducers/developer'
 import {
@@ -24,6 +24,8 @@ import {
 export type AppContentProps = {
   appDetailState: DeveloperAppDetailState
 }
+
+export type CustomUninstallCell = React.FC<{ onClick: () => void }>
 
 interface HandleUninstallSuccessParams {
   handleAfterClose: any
@@ -61,6 +63,38 @@ export const CustomUninstallCell: React.FC<{ onClick: () => void }> = ({ onClick
   <a onClick={onClick}>Uninstall</a>
 )
 
+export const generateInstallationTableColumns = (
+  onUninstall: (app: InstallationModel) => () => void,
+  CustomUninstallCell?: CustomUninstallCell,
+) => () => {
+  const UninstallCell = ({ row }) => {
+    if (CustomUninstallCell) {
+      return <CustomUninstallCell onClick={onUninstall(row.original)} />
+    }
+
+    return (
+      <Button type="button" variant="primary" onClick={onUninstall(row.original)}>
+        Uninstall
+      </Button>
+    )
+  }
+
+  return [
+    {
+      Header: 'Client',
+      accessor: 'client',
+    },
+    {
+      Header: 'Date Installed',
+      accessor: d => dayjs(d.created).format(DATE_TIME_FORMAT.DATE_FORMAT),
+    },
+    {
+      Header: 'Uninstall',
+      Cell: UninstallCell,
+    },
+  ]
+}
+
 const AppContent: React.FC<AppContentProps> = ({ appDetailState }) => {
   const appDetailData = appDetailState.data as AppDetailDataNotNull
   const { summary = '', authFlow = '', externalId = '', id = '', name = '', scopes = [] } = appDetailData
@@ -70,7 +104,10 @@ const AppContent: React.FC<AppContentProps> = ({ appDetailState }) => {
   const developerId = useSelector(selectDeveloperId) || ''
 
   const [uninstallApp, setUninstallApp] = React.useState<InstallationModel>()
-  const columns = generateColumns(handleUninstall(setUninstallApp), CustomUninstallCell)()
+  const installationTableColumns = generateInstallationTableColumns(
+    handleUninstall(setUninstallApp),
+    CustomUninstallCell,
+  )()
   const isVisibleUninstallModal = Boolean(uninstallApp)
 
   return (
@@ -96,7 +133,7 @@ const AppContent: React.FC<AppContentProps> = ({ appDetailState }) => {
       <ListingPreviewSection onClick={handleOpenAppPreview(id)} />
       <AuthenticationSection id={id} authFlow={authFlow} externalId={externalId} />
       <PermissionsSection permissions={scopes} />
-      <InstallationsTableSection columns={columns} data={data} />
+      <InstallationsTableSection columns={installationTableColumns} data={data} />
     </FlexContainerBasic>
   )
 }
